# cmd/api/docs/swagger.yaml
definitions:
  models.Admin:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      statusCode:
        type: integer
    type: object
info:
  contact: {}
paths:
  /admin/login:
    post:
      consumes:
      - application/json
      description: Log in as an admin with provided credentials
      parameters:
      - description: Admin login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Admin'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully logged in
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Provided data is not in required format
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Log in as an admin
      tags:
      - Admin Authentication
  /admin/reported-users:
    get:
      consumes:
      - application/json
      description: Get a list of reported users
      operationId: get-reported-users
      produces:
      - application/json
      responses:
        "200":
          description: successfully showing reported users
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get reported users
      tags:
      - Report Management
  /admin/reported-users/{reportId}:
    get:
      description: Get details of a reported user by ID
      operationId: get-reported-user
      parameters:
      - description: Report ID
        in: path
        name: reportId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: successfully showing the reported user
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: string conversion failed
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get reported user by ID
      tags:
      - Report Management
  /admin/signUp:
    post:
      consumes:
      - application/json
      description: Create a new admin with provided details
      parameters:
      - description: Admin details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Admin'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created admin
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Data given is not in required format
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Create a new admin
      tags:
      - Admin Authentication
...
  /admin/users:
    get:
      consumes:
      - application/json
      description: Retrieve all users
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Number of items per page (default: 3)'
        in: query
        name: count
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved users
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: int conversion failed
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Get all users to admin
      tags:
      - User Management
  /admin/users/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves information about a single user based on the provided
        ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved user information
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad request or invalid ID format
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get a single user by ID
      tags:
      - User Management
    patch:
      consumes:
      - application/json
      description: Block or unblock a user based on the provided ID and block status
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: 'Block status: true to block, false to unblock'
        in: query
        name: block
        required: true
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Successfully blocked/unblocked user
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Boolean conversion failed
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Block or unblock a user
      tags:
      - User Management
...
swagger: "2.0"